modules/stream_segment__analyze/requirements.txt
git+ssh://git@github.com/fortniteclipz/ts_shared.git@master#egg=ts_aws&subdirectory=ts_aws
git+ssh://git@github.com/fortniteclipz/ts_shared.git@master#egg=ts_logger&subdirectory=ts_logger
git+ssh://git@github.com/fortniteclipz/ts_shared.git@master#egg=ts_model&subdirectory=ts_model

opencv-python==3.4.3.18
python-Levenshtein==0.12.0
pytesseract==0.2.4
tensorflow==1.14.0
numpy==1.17.0
scipy==1.3.0
